**BROKEN** Refactored directory structure.
* Broke each dmap record type into its own file for readability
